# packages/shared-types/openapi/hono-api.json
{
   "openapi": "3.0.0",
   "info": {
      "title": "Sonory API",
      "version": "0.1.0",
      "description": "音声ピン管理・AI分析のためのAPI。Hono + Cloudflare Workers上で動作。"
   },
   "servers": [
      {
         "url": "http://localhost:8787",
         "description": "ローカル開発環境"
      }
   ],
...
   "paths": {
...
      "/api/audio/upload-url": {
         "post": {
            "tags": [
               "Audio"
            ],
            "summary": "Presigned URL生成",
            "description": "Supabase Storageへの直接アップロード用Presigned URLを生成",
            "requestBody": {
               "required": true,
               "content": {
                  "application/json": {
                     "schema": {
                        "type": "object",
                        "properties": {
                           "fileName": {
                              "type": "string"
                           },
                           "userId": {
                              "type": "string"
                           }
                        },
                        "required": [
                           "fileName"
                        ]
                     }
                  }
               }
            },
            "responses": {
               "200": {
                  "description": "Presigned URLと関連情報",
                  "content": {
                     "application/json": {
                        "schema": {
                           "type": "object",
                           "properties": {
                              "success": {
                                 "type": "boolean",
                                 "enum": [
                                    true
                                 ]
                              },
                              "data": {
                                 "type": "object",
                                 "properties": {
                                    "uploadUrl": {
                                       "type": "string"
                                    },
                                    "filePath": {
                                       "type": "string"
                                    },
                                    "expiresAt": {
                                       "type": "string"
                                    },
                                    "maxFileSize": {
                                       "type": "number"
                                    }
                                 },
                                 "required": [
                                    "uploadUrl",
                                    "filePath",
                                    "expiresAt",
                                    "maxFileSize"
                                 ]
                              }
                           },
                           "required": [
                              "success",
                              "data"
                           ]
                        }
                     }
                  }
               },
               "400": {
                  "description": "リクエスト不正",
                  "content": {
                     "application/json": {
                        "schema": {
                           "type": "object",
                           "properties": {
                              "success": {
                                 "type": "boolean",
                                 "enum": [
                                    false
                                 ]
                              },
                              "error": {
                                 "type": "object",
                                 "properties": {
                                    "code": {
                                       "type": "string"
                                    },
                                    "message": {
                                       "type": "string"
                                    },
                                    "details": {
                                       "nullable": true
                                    },
                                    "timestamp": {
                                       "type": "string"
                                    },
                                    "requestId": {
                                       "type": "string"
                                    }
                                 },
                                 "required": [
                                    "code",
                                    "message",
                                    "timestamp",
                                    "requestId"
                                 ]
                              }
                           },
                           "required": [
                              "success",
                              "error"
                           ]
                        }
                     }
                  }
               },
               "500": {
                  "description": "サーバーエラー",
                  "content": {
                     "application/json": {
                        "schema": {
                           "type": "object",
                           "properties": {
                              "success": {
                                 "type": "boolean",
                                 "enum": [
                                    false
                                 ]
                              },
                              "error": {
                                 "type": "object",
                                 "properties": {
                                    "code": {
                                       "type": "string"
                                    },
                                    "message": {
                                       "type": "string"
                                    },
                                    "details": {
                                       "nullable": true
                                    },
                                    "timestamp": {
                                       "type": "string"
                                    },
                                    "requestId": {
                                       "type": "string"
                                    }
                                 },
                                 "required": [
                                    "code",
                                    "message",
                                    "timestamp",
                                    "requestId"
                                 ]
                              }
                           },
                           "required": [
                              "success",
                              "error"
                           ]
                        }
                     }
                  }
               }
            }
         }
      },
      "/api/audio/upload": {
         "post": {
            "tags": [
               "Audio"
            ],
            "summary": "音声ファイルアップロード",
            "description": "音声ファイルを直接アップロード（FormData）",
            "requestBody": {
               "required": true,
               "content": {
                  "multipart/form-data": {
                     "schema": {
                        "type": "object",
                        "properties": {
                           "audio": {
                              "nullable": true
                           },
                           "userId": {
                              "type": "string"
                           }
                        }
                     }
                  }
               }
            },
            "responses": {
               "200": {
                  "description": "アップロード結果",
                  "content": {
                     "application/json": {
                        "schema": {
                           "type": "object",
                           "properties": {
                              "success": {
                                 "type": "boolean",
                                 "enum": [
                                    true
                                 ]
                              },
                              "data": {
                                 "type": "object",
                                 "properties": {
                                    "audioId": {
                                       "type": "string"
                                    },
                                    "audioUrl": {
                                       "type": "string"
                                    },
                                    "audioFilePath": {
                                       "type": "string"
                                    },
                                    "metadata": {
                                       "type": "object",
                                       "properties": {
                                          "id": {
                                             "type": "string"
                                          },
                                          "filename": {
                                             "type": "string"
                                          },
                                          "size": {
                                             "type": "number"
                                          },
                                          "format": {
                                             "type": "string",
                                             "enum": [
                                                "webm",
                                                "mp3",
                                                "wav",
                                                "mp4",
                                                "m4a",
                                                "flac",
                                                "ogg"
                                             ]
                                          },
                                          "duration": {
                                             "type": "number"
                                          },
                                          "url": {
                                             "type": "string"
                                          },
                                          "uploadedAt": {
                                             "type": "string"
                                          }
                                       },
                                       "required": [
                                          "id",
                                          "filename",
                                          "size",
                                          "format",
                                          "duration",
                                          "uploadedAt"
                                       ]
                                    }
                                 },
                                 "required": [
                                    "audioId",
                                    "audioUrl",
                                    "audioFilePath",
                                    "metadata"
                                 ]
                              }
                           },
                           "required": [
                              "success",
                              "data"
                           ]
                        }
                     }
                  }
               },
               "400": {
                  "description": "リクエスト不正",
                  "content": {
                     "application/json": {
                        "schema": {
                           "type": "object",
                           "properties": {
                              "success": {
                                 "type": "boolean",
                                 "enum": [
                                    false
                                 ]
                              },
                              "error": {
                                 "type": "object",
                                 "properties": {
                                    "code": {
                                       "type": "string"
                                    },
                                    "message": {
                                       "type": "string"
                                    },
                                    "details": {
                                       "nullable": true
                                    },
                                    "timestamp": {
                                       "type": "string"
                                    },
                                    "requestId": {
                                       "type": "string"
                                    }
                                 },
                                 "required": [
                                    "code",
                                    "message",
                                    "timestamp",
                                    "requestId"
                                 ]
                              }
                           },
                           "required": [
                              "success",
                              "error"
                           ]
                        }
                     }
                  }
               },
               "500": {
                  "description": "サーバーエラー",
                  "content": {
                     "application/json": {
                        "schema": {
                           "type": "object",
                           "properties": {
                              "success": {
                                 "type": "boolean",
                                 "enum": [
                                    false
                                 ]
                              },
                              "error": {
                                 "type": "object",
                                 "properties": {
                                    "code": {
                                       "type": "string"
                                    },
                                    "message": {
                                       "type": "string"
                                    },
                                    "details": {
                                       "nullable": true
                                    },
                                    "timestamp": {
                                       "type": "string"
                                    },
                                    "requestId": {
                                       "type": "string"
                                    }
                                 },
                                 "required": [
                                    "code",
                                    "message",
                                    "timestamp",
                                    "requestId"
                                 ]
                              }
                           },
                           "required": [
                              "success",
                              "error"
                           ]
                        }
                     }
                  }
               }
            }
         }
      },
...
   }
}